chore(index): remove SDK index template; keep shared helpers for component template
Extract expand/flatten and alias helpers into template_sdk_shared.go.
Add template_constants.go for version gates shared with PF package.
Point Plugin Framework version_gating at index package constants.
Remove unused index_template_resource.md embed after SDK resource deletion.

Made-with: Cursor
