--- internal/elasticsearch/index/descriptions/index_template_resource.md
@@ -1,1 +1,0 @@
-Creates or updates an index template. Index templates define settings, mappings, and aliases that can be applied automatically to new indices. See, https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-put-template.html
